requirements.txt.txt
streamlit>=1.28.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0